Update RefreshCurrentPage.lava
This version checks not only whether there is a CampusContext, but also whether one was intended (by checking the URL). This way, if the CampusContext is empty but the URL also didn't specify a CampusId, then it will not go into an infinite loop of reloading the page.

The weakness of this version is that a user could still unintentionally enter an infinite loop pf reloading the page if they change the CampusId number in the URL to a number that doesn't correspond to any CampusId in our database