Rephrase Lava 'assign'
i thought the only way to assign the COUNT from the SQL result was to create a FOR loop for the one row in the SQL result. But i have since learned that i can simply use the `[0]` in order to refer to the first item in the SQL result array.