Update VRL_Internal_Recipes.lava
These tweaks will allow us to use this Lava for the 'ContentChannelItemView' Block rather than the 'HTMLContent' Block.

The only difference is in the way that we assign the variables: The HTMLContent Block required us to use entity commands based on the Id passed in the URL, whereas the ContentChannelItemView Block can pass the CCI object through Page Configurations.