Update Filter Criteria
This DataView is used for auto-removing rows from the [Reservation] Table. 

Previously, it was configured to remove anything where ReservationTypeId=6

We need some Placeholders to remain in the [Reservation] Table. Therefore, i am adjusting the filter criteria such that Placeholders will be removed only if [ApprovalState] is "Draft" or "Cancelled"