QOL update
There was a BlockId-12999; it was a WorkflowEntry Block.

Each time someone interacted with the Workflow in that WorkflowEntry Block, it would refresh PageId 16918 and "jump to the top" of the page.

This behavior was annoying, so i replaced the WorkflowEntry Block with a simple iframe. This way, PageId 16918 does not "jump to the top" when User interacts with the Workflow.